Fixed issue #9 of Crashlytics; Changed data type "int" to "NSInteger" in FCSVenueViewController

+++ FoodCircles/FCSVenueViewController.h
@@ -13,6 +13,6 @@
 
 @interface FCSVenueViewController : FCSStyledViewController
 
-@property int selectedVenueIndex;
+@property NSInteger selectedVenueIndex;
 
 @end
